backend/requirements.txt
# Core Django
Django==5.1.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.3
django-extensions==3.2.3
django-ratelimit==4.1.0


# Database
psycopg2-binary==2.9.11
django-dbbackup==4.0.2

# Real-time & WebSocket
channels==4.1.0
channels-redis==4.1.0
daphne==4.1.2
python-socketio==5.11.0
python-engineio==4.9.0

# Async & Background Tasks
celery==5.4.0
redis==5.0.7
django-redis==5.4.0

# API & Serialization
djangorestframework-simplejwt==5.5.1
coreapi==2.3.3
python-dateutil==2.8.2
pytz==2024.1

# Search & Full-text
django-haystack==3.2.1
elasticsearch==8.14.0

# Development & Utilities
python-dotenv==1.0.1
Pillow==11.0.0
requests==2.31.0

# Testing
pytest==8.0.2
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production
gunicorn==22.0.0
whitenoise==6.7.0

# Monitoring & Logging
sentry-sdk==2.42.1
